# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Projects RBAC API
  description: |
    REST API for authentication, groups, membership, RBAC policies, and protected assets.

    Import this file into Postman (**Import → File**), Insomnia, Thunder Client, or any OpenAPI 3 tool.
  version: 1.0.0
  contact:
    name: Projects API
servers:
  - url: http://localhost:3000
    description: Local unified Next.js app (UI + API)
  - url: "{baseUrl}"
    description: Custom / cloud deployment
    variables:
      baseUrl:
        default: https://app.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdatePermissionsRequest:
      type: object
      required: [scopes, target]
      properties:
        scopes:
          type: array
          items:
            type: string
            enum: [READ, WRITE, DELETE, ADMIN]
          example: [READ, WRITE]
        target:
          type: string
          example: User
        resourceType:
          type: string
          enum: [USER, GROUP, ASSET, DOCUMENT, DASHBOARD, DATASET, SURVEY, SURVEY_RESPONSE]
          default: ASSET
    Permission:
      type: object
      properties:
        _id:
          type: string
        groupId:
          type: string
        resourceType:
          type: string
          enum: [USER, GROUP, ASSET, DOCUMENT, DASHBOARD, DATASET, SURVEY, SURVEY_RESPONSE]
        resourceId:
          type: string
          nullable: true
        target:
          type: string
        permission:
          type: string
          enum: [READ, WRITE, CREATE, DELETE, ADMIN]
...
paths:
...
  /api/permissions:
    get:
      tags: [Permissions]
      summary: List all permissions with principal names
      operationId: listPermissions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All permission rows for management screens
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Permission'
                    - type: object
                      properties:
                        principalType:
                          type: string
                          enum: [USER, GROUP]
                        principalId:
                          type: string
                        principalName:
                          type: string
                        groupName:
                          type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/permissions/catalog:
    get:
      tags: [Permissions]
      summary: Catalog of asset classes and USER/GROUP principals
      operationId: getPermissionCatalog
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Asset classes (only) plus users and groups for ACL editing
  /api/permissions/acl:
    get:
      tags: [Permissions]
      summary: Load Windows-style ACL for selected assets
      operationId: getAssetAcl
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: resourceType
          required: true
          schema:
            type: string
            enum: [DOCUMENT, DASHBOARD, DATASET, SURVEY, SURVEY_RESPONSE]
        - in: query
          name: allObjects
          schema:
            type: boolean
        - in: query
          name: resourceIds
          schema:
            type: string
          description: Comma-separated asset ids
      responses:
        '200':
          description: ACL entries for the selection
    post:
      tags: [Permissions]
      summary: Apply Windows-style ACL to selected assets
      operationId: applyAssetAcl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resourceType, entries]
              properties:
                resourceType:
                  type: string
                  enum: [DOCUMENT, DASHBOARD, DATASET, SURVEY, SURVEY_RESPONSE]
                allObjects:
                  type: boolean
                objects:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      label:
                        type: string
                entries:
                  type: array
                  items:
                    type: object
                    properties:
                      principalType:
                        type: string
                        enum: [USER, GROUP]
                      principalId:
                        type: string
                      scopes:
                        type: array
                        items:
                          type: string
                          enum: [READ, WRITE, CREATE, DELETE, ADMIN]
      responses:
        '200':
          description: ACL applied
        '400':
          description: Validation error
  /api/groups/{groupId}/permissions:
    get:
      tags: [Groups]
      summary: List permissions for a group
      operationId: getGroupPermissions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission rows for the group
          content:
            application/json:
              schema:
                type: object
                properties:
                  groupId:
                    type: string
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '401':
          description: Unauthorized
        '404':
          description: Group not found
    post:
      tags: [Groups]
      summary: Update group permissions
      operationId: updateGroupPermissions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePermissionsRequest'
      responses:
        '200':
          description: Permissions updated in the permissions collection
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Group not found
...